feat(gui): Add GUI to interactively use the design tool
BREAKING CHANGE:

Class variables in mainclass.py are refactored to only indicate
accessible ones being public.

A new config module is integrated to save data for the next startup.

A new option "-g" is available to start the GUI from the command-line.

For installation, setup.py configuration is extended to enable py2app to
generate an OSX application.

Included icons with GPL-3.0 license.

The cgi.html consists of JavaScript written code to pass data to the
cgi_script.py and is included to facilitate web-based CGI usage.

The structure of the mainclass.py and mixin classes is slightly changed
for redundant code removal and better readability.

The README file has been updated accordingly.